+++ central/deployment/service/service.go
@@ -4,7 +4,6 @@ import (
 	"context"
 
 	"github.com/stackrox/rox/central/deployment/datastore"
-	multiplierStore "github.com/stackrox/rox/central/multiplier/store"
 	processIndicatorDataStore "github.com/stackrox/rox/central/processindicator/datastore"
 	processWhitelistDataStore "github.com/stackrox/rox/central/processwhitelist/datastore"
 	processWhitelistResultsStore "github.com/stackrox/rox/central/processwhitelistresults/datastore"
@@ -29,13 +28,12 @@ type Service interface {
 
 // New returns a new Service instance using the given DataStore.
 func New(datastore datastore.DataStore, processIndicators processIndicatorDataStore.DataStore, processWhitelists processWhitelistDataStore.DataStore,
-	processWhitelistResults processWhitelistResultsStore.DataStore, multipliers multiplierStore.Store, manager riskManager.Manager) Service {
+	processWhitelistResults processWhitelistResultsStore.DataStore, manager riskManager.Manager) Service {
 	return &serviceImpl{
 		datastore:               datastore,
 		processIndicators:       processIndicators,
 		processWhitelists:       processWhitelists,
 		processWhitelistResults: processWhitelistResults,
-		multipliers:             multipliers,
 		manager:                 manager,
 	}
 }