+++ central/deployment/service/singleton.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	"github.com/stackrox/rox/central/deployment/datastore"
-	multiplierStore "github.com/stackrox/rox/central/multiplier/store"
 	processIndicatorDataStore "github.com/stackrox/rox/central/processindicator/datastore"
 	processWhitelistDataStore "github.com/stackrox/rox/central/processwhitelist/datastore"
 	processWhitelistResultsStore "github.com/stackrox/rox/central/processwhitelistresults/datastore"
@@ -17,7 +16,7 @@ var (
 )
 
 func initialize() {
-	as = New(datastore.Singleton(), processIndicatorDataStore.Singleton(), processWhitelistDataStore.Singleton(), processWhitelistResultsStore.Singleton(), multiplierStore.Singleton(), manager.Singleton())
+	as = New(datastore.Singleton(), processIndicatorDataStore.Singleton(), processWhitelistDataStore.Singleton(), processWhitelistResultsStore.Singleton(), manager.Singleton())
 }
 
 // Singleton provides the instance of the Service interface to register.