+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project that require documentation updates will be d
 - Queries against time fields involving a duration have now flipped directionality to a more intuitive way.
   Previously, searching `Image Creation Time: >3h` would show all images created _after_ 3 hours before the current time;
   now, it shows all images created more than three hours ago -- that is, _before_ the moment in time 3 hours before the current time.
+- Removed the `/v1/deployments/metadata/multipliers` API.  User defined risk multipliers will no longer be taken into account.
 
 
 ## [23.0]